pyproject.toml
[project]
name = "redisense"
version = "0.1.0"
description = "Smart Energy Usage Monitoring and Anomaly Detection using Redis 8 + AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "redis>=5.0.0",
    "pydantic>=2.7.0",
    "scikit-learn>=1.5.0",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.35.0",
    "plotly>=5.22.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "faker>=25.0.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "click>=8.2.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "fakeredis>=2.24.0",
    "factory-boy>=3.3.0"
]
dev = [
    "black>=24.4.0",
    "flake8>=7.0.0",
    "isort>=5.13.0",
    "mypy>=1.11.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "config"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
    "redis: Tests requiring Redis connection"
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
